Implement fetching of comments from our firebase database
Fetch comments from firebase, format them as a javascript array
stored in a string, then use the chunk templating engine to insert
the string into the webview.
Delete DiscussionActivity and it's associated classes and
resources, because they are no longer needed, i.e. The discussions
are now shown on the question screen when the user presses the
discussion button.