Fix bug preventing comments from being displayed
Add check preventing the parent field of a top-level comment from
being set to the string "null". This happens when converting the
comment Javascript object to JSON, then to a Java Comment object,
a process that is necessary when transfering a comment from the
webview to the McqFragment containing it. The "null" string causes
the jquery-comments plugin to fail with a `Cannot read property
'parent'` error.

Additional changes include:
- Add onErrorReturn handler for fetching comments in MCQFragment
  to handle when there's an error fetching the comments for a
  particular question

- Disable offline persistence for debug versions of the app

- Remove console.log statements from discussion_scripts since they
  are not needed any more

- Remove unnecessary log statements from WebAppInterface

- Remove the bottomBar from TestTakeActivity to create more
  screen-space for displaying comments.