Make commenting system real-time
This commit replaces fetching comments once, when the question page is
loaded, with dynamically inserting and removing them whenever a comment
is added or deleted on the server respectively.
It does so using a Firebase ChildEventListener, that listens for the
addition or removal of a comment from the server, and updates the
jquery-comments plugin accordingly.

A WebviewClient is set on the webview in the question fragment,
so that the ChildEventListener can be instantiated in the
WebviewClient's onPageLoaded callback. As a side-effect, links on the
page do not open with the users external browser anymore. This is
counteracted by adding a shouldOverrideUrlLoading method to the
WebviewClient.